extracted util hooks
