+++ packages/packages/libs/editor-controls/src/controls/html-tag-control.tsx
@@ -9,6 +9,7 @@ import { useBoundProp } from '../bound-prop-context';
 import { ConditionalControlInfotip } from '../components/conditional-control-infotip';
 import ControlActions from '../control-actions/control-actions';
 import { createControl } from '../create-control';
+import { type ControlProps } from '../utils/types';
 
 export type SelectOption = {
 	label: string;
@@ -16,15 +17,22 @@ export type SelectOption = {
 	disabled?: boolean;
 };
 
-type Props = {
+type Props = ControlProps< {
 	options: SelectOption[];
 	onChange?: ( newValue: string | null, previousValue: string | null | undefined ) => void;
 	fallbackLabels?: Record< string, string >;
-};
+} >;
 
 const StyledSelect = styled( Select )( () => ( { '.MuiSelect-select.Mui-disabled': { cursor: 'not-allowed' } } ) );
 
-export const HtmlTagControl = createControl( ( { options, onChange, fallbackLabels = {} }: Props ) => {
+export const HtmlTagControl = createControl( ( props: Props ) => {
+	const {
+		options,
+		onChange,
+		fallbackLabels = {},
+		context: { elementId },
+	} = props;
+
 	const { value, setValue, disabled, placeholder } = useBoundProp( stringPropTypeUtil );
 	const handleChange = ( event: SelectChangeEvent< StringPropValue[ 'value' ] > ) => {
 		const newValue = event.target.value || null;
@@ -33,7 +41,7 @@ export const HtmlTagControl = createControl( ( { options, onChange, fallbackLabe
 		setValue( newValue );
 	};
 
-	const elementLabel = getElementLabel() ?? 'element';
+	const elementLabel = getElementLabel( elementId ) ?? 'element';
 	const infoTipProps = {
 		title: __( 'HTML Tag', 'elementor' ),
 		/* translators: %s is the element name. */
@@ -78,8 +86,8 @@ export const HtmlTagControl = createControl( ( { options, onChange, fallbackLabe
 					disabled={ disabled }
 					fullWidth
 				>
-					{ options.map( ( { label, ...props } ) => (
-						<MenuListItem key={ props.value } { ...props } value={ props.value ?? '' }>
+					{ options.map( ( { label, ...itemProps } ) => (
+						<MenuListItem key={ itemProps.value } { ...itemProps } value={ itemProps.value ?? '' }>
 							{ label }
 						</MenuListItem>
 					) ) }
